tools/requirements.txt
azure-mgmt-resource
azure-mgmt-managementgroups
requests